return empty string if you do not use the service on FetchRIUtilizationPercentage
Previously return float64(0) if you do not use the service on func FetchRIUtilizationPercentage.
But everyone do not need RI utilization of the service that you do not use.
Because it do not need to monitor the RI utilization.

And return []*costexplorer.ReservationCoverageGroup on func FetchRICoveragePercentage.
Previously I ownly defined struct and return it on func FetchRICoveragePercentage.
But it can follow updating aws-sdk-go by using the struct defined in package costexplorer.